Add user and location data storage
We were hardcoding some user data storage before and are now going to
use SQLAlchemy to read and write that information.

The JSON Web Token code is now going to look at the database instead of
the previously hardcoded information.

We still need to hash passwords before we store them.